[WPE][GTK] when flatpak sandbox unavailable, processes are run unsandboxed with no warning https://bugs.webkit.org/show_bug.cgi?id=278578
Reviewed by Carlos Garcia Campos.

If we're running under flatpak, but flatpak-spawn doesn't work, it's
probably better to crash rather than continue to run unsandboxed. This
probably indicates a developer is experimenting, but it could be serious
if it happens by accident for an end user.

* Source/WebKit/UIProcess/Launcher/glib/FlatpakLauncher.cpp:
(WebKit::isFlatpakSpawnUsable):
(WebKit::flatpakSpawn):
* Source/WebKit/UIProcess/Launcher/glib/ProcessLauncherGLib.cpp:
(WebKit::ProcessLauncher::launchProcess):
(WebKit::isFlatpakSpawnUsable): Deleted.

Canonical link: https://commits.webkit.org/283361@main
